platform/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis[hiredis]==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==43.0.1
python-multipart==0.0.12
httpx==0.27.2
celery==5.4.0
aiofiles==24.1.0
structlog==24.4.0
prometheus-client==0.21.0
slowapi==0.1.9
